add locators in many files and add methods
