+++ tests/Personal_account.py
@@ -1,0 +1,103 @@
+from Locators.loginPageLocators import TestLocatorsLoginPageLocators
+from Locators.mainpagelocators import TestLocators
+from selenium import webdriver
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support import expected_conditions
+from selenium.webdriver.support.wait import WebDriverWait
+
+from Locators.profileLocators import TestLocatorsProfileLocators
+
+
+class TestPersonalAccount:
+
+    def test_transfer_to_personal_account(self):
+        # Регистрация
+        user_mail = 'Vitalii_Nesterov_05_622@yandex.ru'
+        user_password = '123456'
+        driver = webdriver.Chrome()
+        driver.get("https://stellarburgers.nomoreparties.site/")
+        # нашел кнопку личный кабинет/нажал на нее
+        driver.find_element(*TestLocators.PERSONAL_ACCOUNT).click()
+        # ожидаю загрузку элемента
+        WebDriverWait(driver, 5).until(expected_conditions.visibility_of_element_located((By.CSS_SELECTOR, ".button_button__33qZ0")))
+        # ввожу данные в поля для входа в аккаунт
+        driver.find_element(*TestLocatorsLoginPageLocators.FIELD_NAME).send_keys(user_mail)
+        driver.find_element(*TestLocatorsLoginPageLocators.FIELD_PASSWORD).send_keys(user_password)
+        driver.find_element(*TestLocatorsLoginPageLocators.BUTTON_ENTER).click()
+        WebDriverWait(driver, 5).until(expected_conditions.element_to_be_clickable(TestLocators.PERSONAL_ACCOUNT))
+        # перехожу в личный кабинет
+        driver.find_element(*TestLocators.PERSONAL_ACCOUNT).click()
+        cur_url = 'https://stellarburgers.nomoreparties.site/account'
+        assert cur_url in driver.current_url
+        driver.quit()
+
+    def test_transfer_to_konstructor_from_personal_account_on_click_konstructor(self):
+        # Регистрация
+        user_mail = 'Vitalii_Nesterov_05_622@yandex.ru'
+        user_password = '123456'
+        driver = webdriver.Chrome()
+        driver.get("https://stellarburgers.nomoreparties.site/")
+        # нашел кнопку личный кабинет/нажал на нее
+        driver.find_element(*TestLocators.PERSONAL_ACCOUNT).click()
+        # ожидаю загрузку элемента
+        WebDriverWait(driver, 5).until(expected_conditions.visibility_of_element_located((By.CSS_SELECTOR, ".button_button__33qZ0")))
+        # ввожу данные в поля для входа в аккаунт
+        driver.find_element(*TestLocatorsLoginPageLocators.FIELD_NAME).send_keys(user_mail)
+        driver.find_element(*TestLocatorsLoginPageLocators.FIELD_PASSWORD).send_keys(user_password)
+        driver.find_element(*TestLocatorsLoginPageLocators.BUTTON_ENTER).click()
+        WebDriverWait(driver, 5).until(expected_conditions.element_to_be_clickable(TestLocators.PERSONAL_ACCOUNT))
+        # перехожу в личный кабинет
+        driver.find_element(*TestLocators.PERSONAL_ACCOUNT).click()
+        # перехожу в конструктор по кнопке конструктор
+        driver.find_element(*TestLocators.CONSTRUCTOR).click()
+        cur_url = 'https://stellarburgers.nomoreparties.site/'
+        assert driver.current_url == cur_url
+        driver.quit()
+
+    def test_transfer_to_konstructor_from_personal_account_on_click_logo(self):
+        # Регистрация
+        user_mail = 'Vitalii_Nesterov_05_622@yandex.ru'
+        user_password = '123456'
+        driver = webdriver.Chrome()
+        driver.get("https://stellarburgers.nomoreparties.site/")
+        # нашел кнопку личный кабинет/нажал на нее
+        driver.find_element(*TestLocators.PERSONAL_ACCOUNT).click()
+        # ожидаю загрузку элемента
+        WebDriverWait(driver, 5).until(
+            expected_conditions.visibility_of_element_located((By.CSS_SELECTOR, ".button_button__33qZ0")))
+        # ввожу данные в поля для входа в аккаунт
+        driver.find_element(*TestLocatorsLoginPageLocators.FIELD_NAME).send_keys(user_mail)
+        driver.find_element(*TestLocatorsLoginPageLocators.FIELD_PASSWORD).send_keys(user_password)
+        driver.find_element(*TestLocatorsLoginPageLocators.BUTTON_ENTER).click()
+        WebDriverWait(driver, 5).until(expected_conditions.element_to_be_clickable(TestLocators.PERSONAL_ACCOUNT))
+        # перехожу в личный кабинет
+        driver.find_element(*TestLocators.PERSONAL_ACCOUNT).click()
+        # перехожу в конструктор по кнопке Logo
+        driver.find_element(*TestLocators.LOGO).click()
+        cur_url = 'https://stellarburgers.nomoreparties.site/'
+        assert driver.current_url == cur_url
+        driver.quit()
+
+    def test_exit_from_account(self):
+        user_mail = 'Vitalii_Nesterov_05_622@yandex.ru'
+        user_password = '123456'
+        driver = webdriver.Chrome()
+        driver.get("https://stellarburgers.nomoreparties.site/")
+        # нашел кнопку личный кабинет/нажал на нее
+        driver.find_element(*TestLocators.PERSONAL_ACCOUNT).click()
+        # ожидаю загрузку элемента
+        WebDriverWait(driver, 5).until(
+            expected_conditions.visibility_of_element_located((By.CSS_SELECTOR, ".button_button__33qZ0")))
+        # ввожу данные в поля для входа в аккаунт
+        driver.find_element(*TestLocatorsLoginPageLocators.FIELD_NAME).send_keys(user_mail)
+        driver.find_element(*TestLocatorsLoginPageLocators.FIELD_PASSWORD).send_keys(user_password)
+        driver.find_element(*TestLocatorsLoginPageLocators.BUTTON_ENTER).click()
+        WebDriverWait(driver, 5).until(expected_conditions.element_to_be_clickable(TestLocatorsLoginPageLocators.BUTTON_ENTER))
+        # перехожу в личный кабинет
+        driver.find_element(*TestLocators.PERSONAL_ACCOUNT).click()
+        WebDriverWait(driver, 5).until(expected_conditions.element_to_be_clickable(TestLocatorsProfileLocators.BUTTON_EXIT))
+        # выхожу из аккаунта
+        driver.find_element(*TestLocatorsProfileLocators.BUTTON_EXIT).click()
+        cur_url = 'https://stellarburgers.nomoreparties.site/account'
+        assert cur_url in driver.current_url
+        driver.quit()
